fix: address latest P0/P1 reviewer feedback
- P0: Fix template RunStatus comparison by calling .to_string()
- P1: Change DeliverPolicy::Last back to ::All for historical data
- P1: Create lib.rs to expose functions for tests
- Refactor main.rs to use library interface

Resolves template compilation and test crate import issues.
